Updated ClassNaming.
