+++ app/src/main/AndroidManifest.xml
@@ -48,7 +48,7 @@
             android:parentActivityName="u.ready_wisc.MenuActivity">
         </activity>
         <activity
-            android:name="edu.parkside.cs.checklist.Checklist_Item_ListView"
+            android:name="edu.parkside.cs.checklist.ChecklistItemListView"
             android:label="@string/title_activity_checklist__item_list_view"
             android:parentActivityName="edu.parkside.cs.checklist.Checklist" >
             <meta-data
@@ -56,23 +56,23 @@
                 android:value="edu.parkside.cs.checklist.Checklist" />
         </activity>
         <activity
-            android:name="edu.parkside.cs.checklist.Checklist_Item_Create"
+            android:name="edu.parkside.cs.checklist.ChecklistItemCreate"
             android:label="@string/title_activity_checklist__item__create"
-            android:parentActivityName="edu.parkside.cs.checklist.Checklist_Item_ListView" >
+            android:parentActivityName="edu.parkside.cs.checklist.ChecklistItemListView" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="edu.parkside.cs.checklist.Checklist_Item_ListView" />
+                android:value="edu.parkside.cs.checklist.ChecklistItemListView" />
         </activity>
         <activity
-            android:name="edu.parkside.cs.checklist.Checklist_Item_Detail"
+            android:name="edu.parkside.cs.checklist.ChecklistItemDetail"
             android:label="@string/title_activity_checklist__item__detail"
-            android:parentActivityName="edu.parkside.cs.checklist.Checklist_Item_ListView" >
+            android:parentActivityName="edu.parkside.cs.checklist.ChecklistItemListView" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="edu.parkside.cs.checklist.Checklist_Item_ListView" />
+                android:value="edu.parkside.cs.checklist.ChecklistItemListView" />
         </activity>
         <activity
-            android:name="edu.parkside.cs.checklist.Checklist_Create"
+            android:name="edu.parkside.cs.checklist.ChecklistCreate"
             android:label="@string/title_activity_checklist__kit__create"
             android:parentActivityName="edu.parkside.cs.checklist.Checklist" >
             <meta-data
@@ -80,12 +80,12 @@
                 android:value="edu.parkside.cs.checklist.Checklist" />
         </activity>
         <activity
-            android:name="edu.parkside.cs.checklist.Checklist_Create_Item_Add"
+            android:name="edu.parkside.cs.checklist.ChecklistCreateItemAdd"
             android:label="@string/title_activity_checklist__create__item__add"
-            android:parentActivityName="edu.parkside.cs.checklist.Checklist_Create" >
+            android:parentActivityName="edu.parkside.cs.checklist.ChecklistCreate" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="edu.parkside.cs.checklist.Checklist_Create" />
+                android:value="edu.parkside.cs.checklist.ChecklistCreate" />
         </activity>
