+++ app/src/main/java/edu/parkside/cs/checklist/ChecklistCreateItemAdapter.java
@@ -202,6 +202,7 @@
  *        limitations under the License.
  */
 
+
 package edu.parkside.cs.checklist;
 
 import android.content.Context;
@@ -209,8 +210,7 @@ import android.content.Intent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ArrayAdapter;
-import android.widget.ProgressBar;
+import android.widget.CheckBox;
 import android.widget.TextView;
 
 import java.util.ArrayList;
@@ -218,28 +218,29 @@ import java.util.ArrayList;
 import u.ready_wisc.R;
 
 /**
- * Maps the activity_checklist.xml to the Checklist_Row objects.
- *
  * @author David Krawchuk
- * @version 1.0v Build * March 18 2015
+ * @version 1.0v Build * March 18 2015 *
+ *          <p/>
+ *          Description:
+ *          The data source for the checklist create item listview.
+ *          Provides the listeners for displaying item details and deletion upon clicking
+ *          the checkbox.
  * @email krawchukdavid@gmail.com
  */
-public class Checklist_ArrayAdapter extends ArrayAdapter {
-
-    /* INSTANCE VARIABLE BLOCK START */
-    Context context;
-    ArrayList<Checklist_Row> list;
-    /* INSTANCE VARIABLE BLOCK END */
+public class ChecklistCreateItemAdapter extends ChecklistItemArrayAdapter {
+    /* Instance variable block. */
+    ArrayList<String> arrayOfDescriptions = null;
+    /* Instance variable block. */
 
     /**
-     * Constructor
+     * Constructor.
      *
      * @param context
      * @param resource
+     * @see ChecklistItemArrayAdapter
      */
-    public Checklist_ArrayAdapter(Context context, int resource) {
+    public ChecklistCreateItemAdapter(Context context, int resource) {
         super(context, resource);
-        this.context = context;
     }
 
     /**
@@ -248,80 +249,58 @@ public class Checklist_ArrayAdapter extends ArrayAdapter {
      * @param context
      * @param resource
      * @param objects
+     * @param arrayOfDescriptions
+     * @see ChecklistItemArrayAdapter
      */
-    public Checklist_ArrayAdapter(Context context, int resource, ArrayList<Checklist_Row> objects) {
+    public ChecklistCreateItemAdapter(Context context, int resource, ArrayList<ChecklistItemRow> objects, ArrayList<String> arrayOfDescriptions) {
         super(context, resource, objects);
-        this.context = context;
-        this.list = objects;
+        this.arrayOfDescriptions = arrayOfDescriptions;
     }
 
     /**
-     * Resets the lists contents and notifies the listener that the adapter's contents have changed.
-     *
-     * @param list
-     */
-    public void setList(ArrayList<Checklist_Row> list) {
-        this.list = list;
-        notifyDataSetChanged();
-    }
-
-    /**
-     * Returns the inflated row filled with the Checklist_Row attributes. Given the current editing
-     * condition allows for selection or deletion of checklist rows.
+     * Provides the data to the source listview that requests the data.
+     * Attaches onclick listeners to the items in order to view item details or delete items when
+     * item checkbox is clicked.
      *
      * @param position
      * @param convertView
      * @param parent
      * @return
-     * @TODO Error condition needs to be implemented!
-     * @TODO Replace string comparisons with index or Object type.
      */
     @Override
-    public View getView(final int position, final View convertView, final ViewGroup parent) {
-        final LayoutInflater inflater = (LayoutInflater) context
-                .getSystemService(Context.LAYOUT_INFLATER_SERVICE);
+    public View getView(final int position, View convertView, ViewGroup parent) {
+        LayoutInflater inflater = (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
 
-        // Layout references.
-        final View rowView = inflater.inflate(R.layout.activity_checklist_row, parent, false);
-        TextView title = (TextView) rowView.findViewById(R.id.activity_checklist_row_textview);
-        ProgressBar progressBar = (ProgressBar) rowView.findViewById(R.id.activity_checklist_row_progressbar);
+        View rowView = inflater.inflate(R.layout.activity_checklist_item_listview_row, parent, false);
 
-        // Apply layout attributes.
-        title.setText(this.list.get(position).getTitle());
-        progressBar.setProgress(this.list.get(position).getProgress());
+        TextView itemName = (TextView) rowView.findViewById(R.id.checklist_item_listview_row_textview);
+        final CheckBox checkBox = (CheckBox) rowView.findViewById(R.id.checklist_item_listview_row_checkbox);
 
-        rowView.setOnClickListener(new View.OnClickListener() {
+        itemName.setText(this.checklist_item_rowArrayList.get(position).getName());
 
+        checkBox.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                if (((Checklist) context).isInEditMode) {
-                    // If the item is not the Add Checklist row, allow removal.
-                    if (position != list.size() - 1) {
-                        int status = Checklist_Contract_Db_Helper.getDb_helper(context).deleteChecklist(list.get(position));
-                        if (status == Checklist_Contract_Db_Helper.FAILURE) {
-                            // Handle error. Possible display an alertView to notify user.
-                        } else {
-                            // Remove row and notify listeners that the adapter contents have changed.
-                            list.remove(position);
-                            notifyDataSetChanged();
-                        }
-                    }
+                // 1. Remove checklist item and description.
+                // 2. Update listview.
+                checklist_item_rowArrayList.remove(position);
+                arrayOfDescriptions.remove(position);
+                notifyDataSetChanged();
+            }
+        });
 
-
-                }
-                // Should use last list index, or some object type instead of string,
-                // bad idea if we end up using localized strings.
-                else if (list.get(position).getTitle().contains("Add Checklist")) {
-                    Intent intent = new Intent(context, Checklist_Create.class);
-                    context.startActivity(intent);
-                } else {
-                    Intent intent = new Intent(context, Checklist_Item_ListView.class);
-                    intent.putExtra(Checklist.EXTRA_MESSAGE, list.get(position));
-                    context.startActivity(intent);
-                }
+        rowView.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(context, ChecklistCreateItemAdd.class);
+                intent.putExtra(ChecklistCreateItemAdd.EXTRA_MESSAGE_ITEM, checklist_item_rowArrayList.get(position));
+                intent.putExtra(ChecklistCreateItemAdd.EXTRA_MESSAGE_DESC, arrayOfDescriptions.get(position));
+                // Send description.
+                ((ChecklistCreate) context).startActivityForResult(intent, ChecklistCreate.UPDATE);
             }
         });
 
         return rowView;
     }
+
 }
