+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/constant-values.html
@@ -117,7 +117,7 @@
         <tbody>
         <tr class="altColor">
             <td class="colFirst"><a
-                    name="edu.parkside.cs.checklist.Checklist_Contract_Db_Helper.DATABASE_NAME">
+                    name="edu.parkside.cs.checklist.ChecklistContractDBHelper.DATABASE_NAME">
                 <!--   -->
             </a><code>public&nbsp;static&nbsp;final&nbsp;java.lang.String</code></td>
             <td><code><a
@@ -127,7 +127,7 @@
         </tr>
         <tr class="rowColor">
             <td class="colFirst"><a
-                    name="edu.parkside.cs.checklist.Checklist_Contract_Db_Helper.DATABASE_VERSION">
+                    name="edu.parkside.cs.checklist.ChecklistContractDBHelper.DATABASE_VERSION">
                 <!--   -->
             </a><code>public&nbsp;static&nbsp;final&nbsp;int</code></td>
             <td><code><a
@@ -137,7 +137,7 @@
         </tr>
         <tr class="altColor">
             <td class="colFirst"><a
-                    name="edu.parkside.cs.checklist.Checklist_Contract_Db_Helper.FAILURE">
+                    name="edu.parkside.cs.checklist.ChecklistContractDBHelper.FAILURE">
                 <!--   -->
             </a><code>public&nbsp;static&nbsp;final&nbsp;int</code></td>
             <td><code><a href="edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html#FAILURE">FAILURE</a></code>
@@ -146,7 +146,7 @@
         </tr>
         <tr class="rowColor">
             <td class="colFirst"><a
-                    name="edu.parkside.cs.checklist.Checklist_Contract_Db_Helper.SUCCESS">
+                    name="edu.parkside.cs.checklist.ChecklistContractDBHelper.SUCCESS">
                 <!--   -->
             </a><code>public&nbsp;static&nbsp;final&nbsp;int</code></td>
             <td><code><a href="edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html#SUCCESS">SUCCESS</a></code>
@@ -429,7 +429,7 @@
         <tbody>
         <tr class="altColor">
             <td class="colFirst"><a
-                    name="edu.parkside.cs.checklist.Checklist_Create.CHECKLIST_DESCRIPTION_ERROR">
+                    name="edu.parkside.cs.checklist.ChecklistCreate.CHECKLIST_DESCRIPTION_ERROR">
                 <!--   -->
             </a><code>static&nbsp;final&nbsp;int</code></td>
             <td><code><a
@@ -439,7 +439,7 @@
         </tr>
         <tr class="rowColor">
             <td class="colFirst"><a
-                    name="edu.parkside.cs.checklist.Checklist_Create.CHECKLIST_INSERT_ERROR">
+                    name="edu.parkside.cs.checklist.ChecklistCreate.CHECKLIST_INSERT_ERROR">
                 <!--   -->
             </a><code>static&nbsp;final&nbsp;int</code></td>
             <td><code><a
@@ -449,7 +449,7 @@
         </tr>
         <tr class="altColor">
             <td class="colFirst"><a
-                    name="edu.parkside.cs.checklist.Checklist_Create.CHECKLIST_ROW_INSERT_ERROR">
+                    name="edu.parkside.cs.checklist.ChecklistCreate.CHECKLIST_ROW_INSERT_ERROR">
                 <!--   -->
             </a><code>static&nbsp;final&nbsp;int</code></td>
             <td><code><a
@@ -458,7 +458,7 @@
             <td class="colLast"><code>2</code></td>
         </tr>
         <tr class="rowColor">
-            <td class="colFirst"><a name="edu.parkside.cs.checklist.Checklist_Create.EXTRA_MESSAGE">
+            <td class="colFirst"><a name="edu.parkside.cs.checklist.ChecklistCreate.EXTRA_MESSAGE">
                 <!--   -->
             </a><code>public&nbsp;static&nbsp;final&nbsp;java.lang.String</code></td>
             <td><code><a href="edu/parkside/cs/checklist/Checklist_Create.html#EXTRA_MESSAGE">EXTRA_MESSAGE</a></code>
@@ -466,7 +466,7 @@
             <td class="colLast"><code>"edu.parkside.cs.checklist_create"</code></td>
         </tr>
         <tr class="altColor">
-            <td class="colFirst"><a name="edu.parkside.cs.checklist.Checklist_Create.NEW">
+            <td class="colFirst"><a name="edu.parkside.cs.checklist.ChecklistCreate.NEW">
                 <!--   -->
             </a><code>static&nbsp;final&nbsp;int</code></td>
             <td><code><a href="edu/parkside/cs/checklist/Checklist_Create.html#NEW">NEW</a></code>
@@ -474,7 +474,7 @@
             <td class="colLast"><code>1</code></td>
         </tr>
         <tr class="rowColor">
-            <td class="colFirst"><a name="edu.parkside.cs.checklist.Checklist_Create.SUCCESS">
+            <td class="colFirst"><a name="edu.parkside.cs.checklist.ChecklistCreate.SUCCESS">
                 <!--   -->
             </a><code>static&nbsp;final&nbsp;int</code></td>
             <td><code><a href="edu/parkside/cs/checklist/Checklist_Create.html#SUCCESS">SUCCESS</a></code>
@@ -482,7 +482,7 @@
             <td class="colLast"><code>0</code></td>
         </tr>
         <tr class="altColor">
-            <td class="colFirst"><a name="edu.parkside.cs.checklist.Checklist_Create.UPDATE">
+            <td class="colFirst"><a name="edu.parkside.cs.checklist.ChecklistCreate.UPDATE">
                 <!--   -->
             </a><code>static&nbsp;final&nbsp;int</code></td>
             <td><code><a
@@ -508,7 +508,7 @@
         <tbody>
         <tr class="altColor">
             <td class="colFirst"><a
-                    name="edu.parkside.cs.checklist.Checklist_Create_Item_Add.EXTRA_MESSAGE_DESC">
+                    name="edu.parkside.cs.checklist.ChecklistCreateItemAdd.EXTRA_MESSAGE_DESC">
                 <!--   -->
             </a><code>public&nbsp;static&nbsp;final&nbsp;java.lang.String</code></td>
             <td><code><a
@@ -518,7 +518,7 @@
         </tr>
         <tr class="rowColor">
             <td class="colFirst"><a
-                    name="edu.parkside.cs.checklist.Checklist_Create_Item_Add.EXTRA_MESSAGE_ITEM">
+                    name="edu.parkside.cs.checklist.ChecklistCreateItemAdd.EXTRA_MESSAGE_ITEM">
                 <!--   -->
             </a><code>public&nbsp;static&nbsp;final&nbsp;java.lang.String</code></td>
             <td><code><a
@@ -544,7 +544,7 @@
         <tbody>
         <tr class="altColor">
             <td class="colFirst"><a
-                    name="edu.parkside.cs.checklist.Checklist_Item_ListView.EXTRA_MESSAGE">
+                    name="edu.parkside.cs.checklist.ChecklistItemListView.EXTRA_MESSAGE">
                 <!--   -->
             </a><code>public&nbsp;static&nbsp;final&nbsp;java.lang.String</code></td>
             <td><code><a
