+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/edu/parkside/cs/checklist/Checklist_ArrayAdapter.html
@@ -100,7 +100,7 @@
                     <li>android.widget.ArrayAdapter</li>
                     <li>
                         <ul class="inheritance">
-                            <li>edu.parkside.cs.checklist.Checklist_ArrayAdapter</li>
+                            <li>edu.parkside.cs.checklist.ChecklistArrayAdapter</li>
                         </ul>
                     </li>
                 </ul>
