+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/edu/parkside/cs/checklist/Checklist_Contract.Checklist_Description_Qureries.html
@@ -183,7 +183,7 @@ extends java.lang.Object</pre>
                             <tr class="altColor">
                                 <td class="colFirst"><code>static java.lang.String</code></td>
                                 <td class="colLast"><code><strong><a
-                                        href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Description_Qureries.html#deleteDescription(edu.parkside.cs.checklist.Checklist_Item_Row)">deleteDescription</a></strong>(<a
+                                        href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Description_Qureries.html#deleteDescription(edu.parkside.cs.checklist.ChecklistItemRow)">deleteDescription</a></strong>(<a
                                         href="../../../../edu/parkside/cs/checklist/Checklist_Item_Row.html"
                                         title="class in edu.parkside.cs.checklist">Checklist_Item_Row</a>&nbsp;item)</code>&nbsp;
                                 </td>
@@ -197,7 +197,7 @@ extends java.lang.Object</pre>
                             <tr class="altColor">
                                 <td class="colFirst"><code>static java.lang.String</code></td>
                                 <td class="colLast"><code><strong><a
-                                        href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Description_Qureries.html#insertDescription(edu.parkside.cs.checklist.Checklist_Item_Row,%20java.lang.String)">insertDescription</a></strong>(<a
+                                        href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Description_Qureries.html#insertDescription(edu.parkside.cs.checklist.ChecklistItemRow,%20java.lang.String)">insertDescription</a></strong>(<a
                                         href="../../../../edu/parkside/cs/checklist/Checklist_Item_Row.html"
                                         title="class in edu.parkside.cs.checklist">Checklist_Item_Row</a>&nbsp;item,
                                     java.lang.String&nbsp;description)</code>&nbsp;</td>
@@ -205,7 +205,7 @@ extends java.lang.Object</pre>
                             <tr class="rowColor">
                                 <td class="colFirst"><code>static java.lang.String</code></td>
                                 <td class="colLast"><code><strong><a
-                                        href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Description_Qureries.html#updateDescription(edu.parkside.cs.checklist.Checklist_Item_Row,%20java.lang.String)">updateDescription</a></strong>(<a
+                                        href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Description_Qureries.html#updateDescription(edu.parkside.cs.checklist.ChecklistItemRow,%20java.lang.String)">updateDescription</a></strong>(<a
                                         href="../../../../edu/parkside/cs/checklist/Checklist_Item_Row.html"
                                         title="class in edu.parkside.cs.checklist">Checklist_Item_Row</a>&nbsp;item,
                                     java.lang.String&nbsp;description)</code>&nbsp;</td>
@@ -287,7 +287,7 @@ extends java.lang.Object</pre>
                                 <pre>public static final&nbsp;java.lang.String&nbsp;getDescriptionWithItemEntryID(int&nbsp;entryID)</pre>
                             </li>
                         </ul>
-                        <a name="insertDescription(edu.parkside.cs.checklist.Checklist_Item_Row, java.lang.String)">
+                        <a name="insertDescription(edu.parkside.cs.checklist.ChecklistItemRow, java.lang.String)">
                             <!--   -->
                         </a>
                         <ul class="blockList">
@@ -299,7 +299,7 @@ extends java.lang.Object</pre>
                                  java.lang.String&nbsp;description)</pre>
                             </li>
                         </ul>
-                        <a name="updateDescription(edu.parkside.cs.checklist.Checklist_Item_Row, java.lang.String)">
+                        <a name="updateDescription(edu.parkside.cs.checklist.ChecklistItemRow, java.lang.String)">
                             <!--   -->
                         </a>
                         <ul class="blockList">
