+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html
@@ -190,7 +190,7 @@ extends java.lang.Object</pre>
                         <tr class="altColor">
                             <td class="colFirst"><code>static java.lang.String[]</code></td>
                             <td class="colLast"><code><strong><a
-                                    href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#deleteItem(edu.parkside.cs.checklist.Checklist_Item_Row)">deleteItem</a></strong>(<a
+                                    href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#deleteItem(edu.parkside.cs.checklist.ChecklistItemRow)">deleteItem</a></strong>(<a
                                     href="../../../../edu/parkside/cs/checklist/Checklist_Item_Row.html"
                                     title="class in edu.parkside.cs.checklist">Checklist_Item_Row</a>&nbsp;item)</code>&nbsp;
                             </td>
@@ -198,7 +198,7 @@ extends java.lang.Object</pre>
                         <tr class="rowColor">
                             <td class="colFirst"><code>static java.lang.String</code></td>
                             <td class="colLast"><code><strong><a
-                                    href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#fetchItem(edu.parkside.cs.checklist.Checklist_Item_Row)">fetchItem</a></strong>(<a
+                                    href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#fetchItem(edu.parkside.cs.checklist.ChecklistItemRow)">fetchItem</a></strong>(<a
                                     href="../../../../edu/parkside/cs/checklist/Checklist_Item_Row.html"
                                     title="class in edu.parkside.cs.checklist">Checklist_Item_Row</a>&nbsp;item)</code>&nbsp;
                             </td>
@@ -206,7 +206,7 @@ extends java.lang.Object</pre>
                         <tr class="altColor">
                             <td class="colFirst"><code>static java.lang.String</code></td>
                             <td class="colLast"><code><strong><a
-                                    href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#fetchItems(edu.parkside.cs.checklist.Checklist_Row)">fetchItems</a></strong>(<a
+                                    href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#fetchItems(edu.parkside.cs.checklist.ChecklistRow)">fetchItems</a></strong>(<a
                                     href="../../../../edu/parkside/cs/checklist/Checklist_Row.html"
                                     title="class in edu.parkside.cs.checklist">Checklist_Row</a>&nbsp;checklist)</code>&nbsp;
                             </td>
@@ -214,7 +214,7 @@ extends java.lang.Object</pre>
                         <tr class="rowColor">
                             <td class="colFirst"><code>static java.lang.String</code></td>
                             <td class="colLast"><code><strong><a
-                                    href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#insertItem(edu.parkside.cs.checklist.Checklist_Item_Row)">insertItem</a></strong>(<a
+                                    href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#insertItem(edu.parkside.cs.checklist.ChecklistItemRow)">insertItem</a></strong>(<a
                                     href="../../../../edu/parkside/cs/checklist/Checklist_Item_Row.html"
                                     title="class in edu.parkside.cs.checklist">Checklist_Item_Row</a>&nbsp;item)</code>&nbsp;
                             </td>
@@ -222,7 +222,7 @@ extends java.lang.Object</pre>
                         <tr class="altColor">
                             <td class="colFirst"><code>static java.lang.String</code></td>
                             <td class="colLast"><code><strong><a
-                                    href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#updateItem(edu.parkside.cs.checklist.Checklist_Item_Row)">updateItem</a></strong>(<a
+                                    href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#updateItem(edu.parkside.cs.checklist.ChecklistItemRow)">updateItem</a></strong>(<a
                                     href="../../../../edu/parkside/cs/checklist/Checklist_Item_Row.html"
                                     title="class in edu.parkside.cs.checklist">Checklist_Item_Row</a>&nbsp;item)</code>&nbsp;
                             </td>
