+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html
@@ -187,7 +187,7 @@ extends java.lang.Object</pre>
                         <tr class="altColor">
                             <td class="colFirst"><code>static java.lang.String</code></td>
                             <td class="colLast"><code><strong><a
-                                    href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html#deleteChecklist(edu.parkside.cs.checklist.Checklist_Row)">deleteChecklist</a></strong>(<a
+                                    href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html#deleteChecklist(edu.parkside.cs.checklist.ChecklistRow)">deleteChecklist</a></strong>(<a
                                     href="../../../../edu/parkside/cs/checklist/Checklist_Row.html"
                                     title="class in edu.parkside.cs.checklist">Checklist_Row</a>&nbsp;checklist_row)</code>&nbsp;
                             </td>
@@ -195,7 +195,7 @@ extends java.lang.Object</pre>
                         <tr class="rowColor">
                             <td class="colFirst"><code>static java.lang.String</code></td>
                             <td class="colLast"><code><strong><a
-                                    href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html#fetchChecklist(edu.parkside.cs.checklist.Checklist_Row)">fetchChecklist</a></strong>(<a
+                                    href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html#fetchChecklist(edu.parkside.cs.checklist.ChecklistRow)">fetchChecklist</a></strong>(<a
                                     href="../../../../edu/parkside/cs/checklist/Checklist_Row.html"
                                     title="class in edu.parkside.cs.checklist">Checklist_Row</a>&nbsp;checklist_row)</code>&nbsp;
                             </td>
@@ -203,7 +203,7 @@ extends java.lang.Object</pre>
                         <tr class="altColor">
                             <td class="colFirst"><code>static java.lang.String</code></td>
                             <td class="colLast"><code><strong><a
-                                    href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html#insertChecklist(edu.parkside.cs.checklist.Checklist_Row)">insertChecklist</a></strong>(<a
+                                    href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html#insertChecklist(edu.parkside.cs.checklist.ChecklistRow)">insertChecklist</a></strong>(<a
                                     href="../../../../edu/parkside/cs/checklist/Checklist_Row.html"
                                     title="class in edu.parkside.cs.checklist">Checklist_Row</a>&nbsp;checklist_row)</code>&nbsp;
                             </td>
@@ -211,7 +211,7 @@ extends java.lang.Object</pre>
                         <tr class="rowColor">
                             <td class="colFirst"><code>static java.lang.String</code></td>
                             <td class="colLast"><code><strong><a
-                                    href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html#updateChecklist(edu.parkside.cs.checklist.Checklist_Row)">updateChecklist</a></strong>(<a
+                                    href="../../../../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html#updateChecklist(edu.parkside.cs.checklist.ChecklistRow)">updateChecklist</a></strong>(<a
                                     href="../../../../edu/parkside/cs/checklist/Checklist_Row.html"
                                     title="class in edu.parkside.cs.checklist">Checklist_Row</a>&nbsp;checklist_row)</code>&nbsp;
                             </td>
