+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/edu/parkside/cs/checklist/Checklist_Create.html
@@ -115,7 +115,7 @@
                                                     <li>
                                                         <ul class="inheritance">
                                                             <li>
-                                                                edu.parkside.cs.checklist.Checklist_Create
+                                                                edu.parkside.cs.checklist.ChecklistCreate
                                                             </li>
                                                         </ul>
                                                     </li>
@@ -781,7 +781,7 @@ extends android.support.v7.app.ActionBarActivity</pre>
                 <dl>
                     <dt><span class="strong">See Also:</span></dt>
                     <dd>
-                        <a href="../../../../constant-values.html#edu.parkside.cs.checklist.Checklist_Create.EXTRA_MESSAGE">Constant
+                        <a href="../../../../constant-values.html#edu.parkside.cs.checklist.ChecklistCreate.EXTRA_MESSAGE">Constant
                             Field Values</a></dd>
                 </dl>
             </li>
@@ -796,7 +796,7 @@ extends android.support.v7.app.ActionBarActivity</pre>
                 <dl>
                     <dt><span class="strong">See Also:</span></dt>
                     <dd>
-                        <a href="../../../../constant-values.html#edu.parkside.cs.checklist.Checklist_Create.NEW">Constant
+                        <a href="../../../../constant-values.html#edu.parkside.cs.checklist.ChecklistCreate.NEW">Constant
                             Field Values</a></dd>
                 </dl>
             </li>
@@ -811,7 +811,7 @@ extends android.support.v7.app.ActionBarActivity</pre>
                 <dl>
                     <dt><span class="strong">See Also:</span></dt>
                     <dd>
-                        <a href="../../../../constant-values.html#edu.parkside.cs.checklist.Checklist_Create.UPDATE">Constant
+                        <a href="../../../../constant-values.html#edu.parkside.cs.checklist.ChecklistCreate.UPDATE">Constant
                             Field Values</a></dd>
                 </dl>
             </li>
@@ -826,7 +826,7 @@ extends android.support.v7.app.ActionBarActivity</pre>
                 <dl>
                     <dt><span class="strong">See Also:</span></dt>
                     <dd>
-                        <a href="../../../../constant-values.html#edu.parkside.cs.checklist.Checklist_Create.SUCCESS">Constant
+                        <a href="../../../../constant-values.html#edu.parkside.cs.checklist.ChecklistCreate.SUCCESS">Constant
                             Field Values</a></dd>
                 </dl>
             </li>
@@ -841,7 +841,7 @@ extends android.support.v7.app.ActionBarActivity</pre>
                 <dl>
                     <dt><span class="strong">See Also:</span></dt>
                     <dd>
-                        <a href="../../../../constant-values.html#edu.parkside.cs.checklist.Checklist_Create.CHECKLIST_INSERT_ERROR">Constant
+                        <a href="../../../../constant-values.html#edu.parkside.cs.checklist.ChecklistCreate.CHECKLIST_INSERT_ERROR">Constant
                             Field Values</a></dd>
                 </dl>
             </li>
@@ -856,7 +856,7 @@ extends android.support.v7.app.ActionBarActivity</pre>
                 <dl>
                     <dt><span class="strong">See Also:</span></dt>
                     <dd>
-                        <a href="../../../../constant-values.html#edu.parkside.cs.checklist.Checklist_Create.CHECKLIST_ROW_INSERT_ERROR">Constant
+                        <a href="../../../../constant-values.html#edu.parkside.cs.checklist.ChecklistCreate.CHECKLIST_ROW_INSERT_ERROR">Constant
                             Field Values</a></dd>
                 </dl>
             </li>
@@ -871,7 +871,7 @@ extends android.support.v7.app.ActionBarActivity</pre>
                 <dl>
                     <dt><span class="strong">See Also:</span></dt>
                     <dd>
-                        <a href="../../../../constant-values.html#edu.parkside.cs.checklist.Checklist_Create.CHECKLIST_DESCRIPTION_ERROR">Constant
+                        <a href="../../../../constant-values.html#edu.parkside.cs.checklist.ChecklistCreate.CHECKLIST_DESCRIPTION_ERROR">Constant
                             Field Values</a></dd>
                 </dl>
             </li>
