+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/edu/parkside/cs/checklist/Checklist_Create_Item_Adapter.html
@@ -107,11 +107,11 @@
                         <ul class="inheritance">
                             <li>
                                 <a href="../../../../edu/parkside/cs/checklist/Checklist_Item_ArrayAdapter.html"
-                                   title="class in edu.parkside.cs.checklist">edu.parkside.cs.checklist.Checklist_Item_ArrayAdapter</a>
+                                   title="class in edu.parkside.cs.checklist">edu.parkside.cs.checklist.ChecklistItemArrayAdapter</a>
                             </li>
                             <li>
                                 <ul class="inheritance">
-                                    <li>edu.parkside.cs.checklist.Checklist_Create_Item_Adapter</li>
+                                    <li>edu.parkside.cs.checklist.ChecklistCreateItemAdapter</li>
                                 </ul>
                             </li>
                         </ul>
@@ -177,7 +177,7 @@ extends <a href="../../../../edu/parkside/cs/checklist/Checklist_Item_ArrayAdapt
                     </table>
                     <ul class="blockList">
                         <li class="blockList"><a
-                                name="fields_inherited_from_class_edu.parkside.cs.checklist.Checklist_Item_ArrayAdapter">
+                                name="fields_inherited_from_class_edu.parkside.cs.checklist.ChecklistItemArrayAdapter">
                             <!--   -->
                         </a>
 
@@ -268,7 +268,7 @@ extends <a href="../../../../edu/parkside/cs/checklist/Checklist_Item_ArrayAdapt
                     </table>
                     <ul class="blockList">
                         <li class="blockList"><a
-                                name="methods_inherited_from_class_edu.parkside.cs.checklist.Checklist_Item_ArrayAdapter">
+                                name="methods_inherited_from_class_edu.parkside.cs.checklist.ChecklistItemArrayAdapter">
                             <!--   -->
                         </a>
