+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/edu/parkside/cs/checklist/Checklist_Create_Item_Add.html
@@ -117,12 +117,12 @@
                                                         <ul class="inheritance">
                                                             <li>
                                                                 <a href="../../../../edu/parkside/cs/checklist/Checklist_Item_Create.html"
-                                                                   title="class in edu.parkside.cs.checklist">edu.parkside.cs.checklist.Checklist_Item_Create</a>
+                                                                   title="class in edu.parkside.cs.checklist">edu.parkside.cs.checklist.ChecklistItemCreate</a>
                                                             </li>
                                                             <li>
                                                                 <ul class="inheritance">
                                                                     <li>
-                                                                        edu.parkside.cs.checklist.Checklist_Create_Item_Add
+                                                                        edu.parkside.cs.checklist.ChecklistCreateItemAdd
                                                                     </li>
                                                                 </ul>
                                                             </li>
@@ -251,7 +251,7 @@ extends <a href="../../../../edu/parkside/cs/checklist/Checklist_Item_Create.htm
         </table>
         <ul class="blockList">
             <li class="blockList"><a
-                    name="fields_inherited_from_class_edu.parkside.cs.checklist.Checklist_Item_Create">
+                    name="fields_inherited_from_class_edu.parkside.cs.checklist.ChecklistItemCreate">
                 <!--   -->
             </a>
 
@@ -396,7 +396,7 @@ extends <a href="../../../../edu/parkside/cs/checklist/Checklist_Item_Create.htm
         </table>
         <ul class="blockList">
             <li class="blockList"><a
-                    name="methods_inherited_from_class_edu.parkside.cs.checklist.Checklist_Item_Create">
+                    name="methods_inherited_from_class_edu.parkside.cs.checklist.ChecklistItemCreate">
                 <!--   -->
             </a>
 
@@ -600,7 +600,7 @@ extends <a href="../../../../edu/parkside/cs/checklist/Checklist_Item_Create.htm
                             <dl>
                                 <dt><span class="strong">See Also:</span></dt>
                                 <dd>
-                                    <a href="../../../../constant-values.html#edu.parkside.cs.checklist.Checklist_Create_Item_Add.EXTRA_MESSAGE_ITEM">Constant
+                                    <a href="../../../../constant-values.html#edu.parkside.cs.checklist.ChecklistCreateItemAdd.EXTRA_MESSAGE_ITEM">Constant
                                         Field Values</a></dd>
                             </dl>
                         </li>
@@ -615,7 +615,7 @@ extends <a href="../../../../edu/parkside/cs/checklist/Checklist_Item_Create.htm
                             <dl>
                                 <dt><span class="strong">See Also:</span></dt>
                                 <dd>
-                                    <a href="../../../../constant-values.html#edu.parkside.cs.checklist.Checklist_Create_Item_Add.EXTRA_MESSAGE_DESC">Constant
+                                    <a href="../../../../constant-values.html#edu.parkside.cs.checklist.ChecklistCreateItemAdd.EXTRA_MESSAGE_DESC">Constant
                                         Field Values</a></dd>
                             </dl>
                         </li>
