+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/edu/parkside/cs/checklist/Checklist_Item_ArrayAdapter.html
@@ -105,7 +105,7 @@
                     </li>
                     <li>
                         <ul class="inheritance">
-                            <li>edu.parkside.cs.checklist.Checklist_Item_ArrayAdapter</li>
+                            <li>edu.parkside.cs.checklist.ChecklistItemArrayAdapter</li>
                         </ul>
                     </li>
                 </ul>
