+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/edu/parkside/cs/checklist/Checklist_Item_Create.html
@@ -115,7 +115,7 @@
                                                     <li>
                                                         <ul class="inheritance">
                                                             <li>
-                                                                edu.parkside.cs.checklist.Checklist_Item_Create
+                                                                edu.parkside.cs.checklist.ChecklistItemCreate
                                                             </li>
                                                         </ul>
                                                     </li>
