+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/edu/parkside/cs/checklist/Checklist_Item_Detail.html
@@ -115,7 +115,7 @@
                                                     <li>
                                                         <ul class="inheritance">
                                                             <li>
-                                                                edu.parkside.cs.checklist.Checklist_Item_Detail
+                                                                edu.parkside.cs.checklist.ChecklistItemDetail
                                                             </li>
                                                         </ul>
                                                     </li>
