+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/edu/parkside/cs/checklist/Checklist_Item_ListView.html
@@ -115,7 +115,7 @@
                                                     <li>
                                                         <ul class="inheritance">
                                                             <li>
-                                                                edu.parkside.cs.checklist.Checklist_Item_ListView
+                                                                edu.parkside.cs.checklist.ChecklistItemListView
                                                             </li>
                                                         </ul>
                                                     </li>
@@ -636,7 +636,7 @@ extends android.support.v7.app.ActionBarActivity</pre>
                 <dl>
                     <dt><span class="strong">See Also:</span></dt>
                     <dd>
-                        <a href="../../../../constant-values.html#edu.parkside.cs.checklist.Checklist_Item_ListView.EXTRA_MESSAGE">Constant
+                        <a href="../../../../constant-values.html#edu.parkside.cs.checklist.ChecklistItemListView.EXTRA_MESSAGE">Constant
                             Field Values</a></dd>
                 </dl>
             </li>
