+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/edu/parkside/cs/checklist/Checklist_Item_Row.html
@@ -94,7 +94,7 @@
     <li>java.lang.Object</li>
     <li>
         <ul class="inheritance">
-            <li>edu.parkside.cs.checklist.Checklist_Item_Row</li>
+            <li>edu.parkside.cs.checklist.ChecklistItemRow</li>
         </ul>
     </li>
 </ul>
