+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/edu/parkside/cs/checklist/Checklist_Row.html
@@ -92,7 +92,7 @@
     <li>java.lang.Object</li>
     <li>
         <ul class="inheritance">
-            <li>edu.parkside.cs.checklist.Checklist_Row</li>
+            <li>edu.parkside.cs.checklist.ChecklistRow</li>
         </ul>
     </li>
 </ul>
