+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/edu/parkside/cs/checklist/class-use/Checklist_Create_Item_Add.html
@@ -4,7 +4,7 @@
 <html lang="en">
 <head>
     <!-- Generated by javadoc (version 1.7.0_75) on Wed Mar 18 20:30:10 CDT 2015 -->
-    <title>Uses of Class edu.parkside.cs.checklist.Checklist_Create_Item_Add</title>
+    <title>Uses of Class edu.parkside.cs.checklist.ChecklistCreateItemAdd</title>
     <meta name="date" content="2015-03-18">
     <link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
 </head>
@@ -68,10 +68,10 @@
     </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-    <h2 title="Uses of Class edu.parkside.cs.checklist.Checklist_Create_Item_Add" class="title">Uses
-        of Class<br>edu.parkside.cs.checklist.Checklist_Create_Item_Add</h2>
+    <h2 title="Uses of Class edu.parkside.cs.checklist.ChecklistCreateItemAdd" class="title">Uses
+        of Class<br>edu.parkside.cs.checklist.ChecklistCreateItemAdd</h2>
 </div>
-<div class="classUseContainer">No usage of edu.parkside.cs.checklist.Checklist_Create_Item_Add</div>
+<div class="classUseContainer">No usage of edu.parkside.cs.checklist.ChecklistCreateItemAdd</div>
 <!-- ======= START OF BOTTOM NAVBAR ====== -->
 <div class="bottomNav"><a name="navbar_bottom">
     <!--   -->
