+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/edu/parkside/cs/checklist/class-use/Checklist_Item_ArrayAdapter.html
@@ -4,7 +4,7 @@
 <html lang="en">
 <head>
     <!-- Generated by javadoc (version 1.7.0_75) on Wed Mar 18 20:30:10 CDT 2015 -->
-    <title>Uses of Class edu.parkside.cs.checklist.Checklist_Item_ArrayAdapter</title>
+    <title>Uses of Class edu.parkside.cs.checklist.ChecklistItemArrayAdapter</title>
     <meta name="date" content="2015-03-18">
     <link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
 </head>
@@ -68,8 +68,8 @@
     </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-    <h2 title="Uses of Class edu.parkside.cs.checklist.Checklist_Item_ArrayAdapter" class="title">
-        Uses of Class<br>edu.parkside.cs.checklist.Checklist_Item_ArrayAdapter</h2>
+    <h2 title="Uses of Class edu.parkside.cs.checklist.ChecklistItemArrayAdapter" class="title">
+        Uses of Class<br>edu.parkside.cs.checklist.ChecklistItemArrayAdapter</h2>
 </div>
 <div class="classUseContainer">
     <ul class="blockList">
