+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/edu/parkside/cs/checklist/class-use/Checklist_Item_Detail.html
@@ -4,7 +4,7 @@
 <html lang="en">
 <head>
     <!-- Generated by javadoc (version 1.7.0_75) on Wed Mar 18 20:30:09 CDT 2015 -->
-    <title>Uses of Class edu.parkside.cs.checklist.Checklist_Item_Detail</title>
+    <title>Uses of Class edu.parkside.cs.checklist.ChecklistItemDetail</title>
     <meta name="date" content="2015-03-18">
     <link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
 </head>
@@ -68,10 +68,10 @@
     </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-    <h2 title="Uses of Class edu.parkside.cs.checklist.Checklist_Item_Detail" class="title">Uses of
-        Class<br>edu.parkside.cs.checklist.Checklist_Item_Detail</h2>
+    <h2 title="Uses of Class edu.parkside.cs.checklist.ChecklistItemDetail" class="title">Uses of
+        Class<br>edu.parkside.cs.checklist.ChecklistItemDetail</h2>
 </div>
-<div class="classUseContainer">No usage of edu.parkside.cs.checklist.Checklist_Item_Detail</div>
+<div class="classUseContainer">No usage of edu.parkside.cs.checklist.ChecklistItemDetail</div>
 <!-- ======= START OF BOTTOM NAVBAR ====== -->
 <div class="bottomNav"><a name="navbar_bottom">
     <!--   -->
