+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/index-files/index-1.html
@@ -89,7 +89,7 @@
             </div>
         </dd>
         <dt><span class="strong"><a
-                href="../edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html#addChecklist(edu.parkside.cs.checklist.Checklist_Row)">addChecklist(Checklist_Row)</a></span>
+                href="../edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html#addChecklist(edu.parkside.cs.checklist.ChecklistRow)">addChecklist(Checklist_Row)</a></span>
             - Method in class edu.parkside.cs.checklist.<a
                     href="../edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html"
                     title="class in edu.parkside.cs.checklist">Checklist_Contract_Db_Helper</a></dt>
