+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/index-files/index-14.html
@@ -117,7 +117,7 @@
             <div class="block">Takes an array of queries.</div>
         </dd>
         <dt><span class="strong"><a
-                href="../edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html#returnDescriptionFromItem(edu.parkside.cs.checklist.Checklist_Item_Row)">returnDescriptionFromItem(Checklist_Item_Row)</a></span>
+                href="../edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html#returnDescriptionFromItem(edu.parkside.cs.checklist.ChecklistItemRow)">returnDescriptionFromItem(Checklist_Item_Row)</a></span>
             - Method in class edu.parkside.cs.checklist.<a
                     href="../edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html"
                     title="class in edu.parkside.cs.checklist">Checklist_Contract_Db_Helper</a></dt>
