+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/index-files/index-17.html
@@ -93,14 +93,14 @@
             <div class="block">Called by the activity when the user presses the update button.</div>
         </dd>
         <dt><span class="strong"><a
-                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html#updateChecklist(edu.parkside.cs.checklist.Checklist_Row)">updateChecklist(Checklist_Row)</a></span>
+                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html#updateChecklist(edu.parkside.cs.checklist.ChecklistRow)">updateChecklist(Checklist_Row)</a></span>
             - Static method in class edu.parkside.cs.checklist.<a
                     href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html"
                     title="class in edu.parkside.cs.checklist">Checklist_Contract.Checklist_Queries</a>
         </dt>
         <dd>&nbsp;</dd>
         <dt><span class="strong"><a
-                href="../edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html#updateChecklist(edu.parkside.cs.checklist.Checklist_Row)">updateChecklist(Checklist_Row)</a></span>
+                href="../edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html#updateChecklist(edu.parkside.cs.checklist.ChecklistRow)">updateChecklist(Checklist_Row)</a></span>
             - Method in class edu.parkside.cs.checklist.<a
                     href="../edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html"
                     title="class in edu.parkside.cs.checklist">Checklist_Contract_Db_Helper</a></dt>
@@ -118,21 +118,21 @@
             <div class="block">Updates the checklist percentage attribute.</div>
         </dd>
         <dt><span class="strong"><a
-                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Description_Qureries.html#updateDescription(edu.parkside.cs.checklist.Checklist_Item_Row,%20java.lang.String)">updateDescription(Checklist_Item_Row,
+                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Description_Qureries.html#updateDescription(edu.parkside.cs.checklist.ChecklistItemRow,%20java.lang.String)">updateDescription(Checklist_Item_Row,
             String)</a></span> - Static method in class edu.parkside.cs.checklist.<a
                 href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Description_Qureries.html"
                 title="class in edu.parkside.cs.checklist">Checklist_Contract.Checklist_Description_Qureries</a>
         </dt>
         <dd>&nbsp;</dd>
         <dt><span class="strong"><a
-                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#updateItem(edu.parkside.cs.checklist.Checklist_Item_Row)">updateItem(Checklist_Item_Row)</a></span>
+                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#updateItem(edu.parkside.cs.checklist.ChecklistItemRow)">updateItem(Checklist_Item_Row)</a></span>
             - Static method in class edu.parkside.cs.checklist.<a
                     href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html"
                     title="class in edu.parkside.cs.checklist">Checklist_Contract.Checklist_Item_Queries</a>
         </dt>
         <dd>&nbsp;</dd>
         <dt><span class="strong"><a
-                href="../edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html#updateItem(edu.parkside.cs.checklist.Checklist_Item_Row,%20java.lang.String)">updateItem(Checklist_Item_Row,
+                href="../edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html#updateItem(edu.parkside.cs.checklist.ChecklistItemRow,%20java.lang.String)">updateItem(Checklist_Item_Row,
             String)</a></span> - Method in class edu.parkside.cs.checklist.<a
                 href="../edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html"
                 title="class in edu.parkside.cs.checklist">Checklist_Contract_Db_Helper</a></dt>
