+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/index-files/index-3.html
@@ -97,14 +97,14 @@
                     title="class in edu.parkside.cs.checklist">Checklist_Contract_Db_Helper</a></dt>
         <dd>&nbsp;</dd>
         <dt><span class="strong"><a
-                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html#deleteChecklist(edu.parkside.cs.checklist.Checklist_Row)">deleteChecklist(Checklist_Row)</a></span>
+                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html#deleteChecklist(edu.parkside.cs.checklist.ChecklistRow)">deleteChecklist(Checklist_Row)</a></span>
             - Static method in class edu.parkside.cs.checklist.<a
                     href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html"
                     title="class in edu.parkside.cs.checklist">Checklist_Contract.Checklist_Queries</a>
         </dt>
         <dd>&nbsp;</dd>
         <dt><span class="strong"><a
-                href="../edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html#deleteChecklist(edu.parkside.cs.checklist.Checklist_Row)">deleteChecklist(Checklist_Row)</a></span>
+                href="../edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html#deleteChecklist(edu.parkside.cs.checklist.ChecklistRow)">deleteChecklist(Checklist_Row)</a></span>
             - Method in class edu.parkside.cs.checklist.<a
                     href="../edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html"
                     title="class in edu.parkside.cs.checklist">Checklist_Contract_Db_Helper</a></dt>
@@ -112,14 +112,14 @@
             <div class="block">Removes the provided checklist row from the Checklist table.</div>
         </dd>
         <dt><span class="strong"><a
-                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Description_Qureries.html#deleteDescription(edu.parkside.cs.checklist.Checklist_Item_Row)">deleteDescription(Checklist_Item_Row)</a></span>
+                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Description_Qureries.html#deleteDescription(edu.parkside.cs.checklist.ChecklistItemRow)">deleteDescription(Checklist_Item_Row)</a></span>
             - Static method in class edu.parkside.cs.checklist.<a
                     href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Description_Qureries.html"
                     title="class in edu.parkside.cs.checklist">Checklist_Contract.Checklist_Description_Qureries</a>
         </dt>
         <dd>&nbsp;</dd>
         <dt><span class="strong"><a
-                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#deleteItem(edu.parkside.cs.checklist.Checklist_Item_Row)">deleteItem(Checklist_Item_Row)</a></span>
+                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#deleteItem(edu.parkside.cs.checklist.ChecklistItemRow)">deleteItem(Checklist_Item_Row)</a></span>
             - Static method in class edu.parkside.cs.checklist.<a
                     href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html"
                     title="class in edu.parkside.cs.checklist">Checklist_Contract.Checklist_Item_Queries</a>
