+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/index-files/index-5.html
@@ -85,21 +85,21 @@
                     title="class in edu.parkside.cs.checklist">Checklist_Contract_Db_Helper</a></dt>
         <dd>&nbsp;</dd>
         <dt><span class="strong"><a
-                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html#fetchChecklist(edu.parkside.cs.checklist.Checklist_Row)">fetchChecklist(Checklist_Row)</a></span>
+                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html#fetchChecklist(edu.parkside.cs.checklist.ChecklistRow)">fetchChecklist(Checklist_Row)</a></span>
             - Static method in class edu.parkside.cs.checklist.<a
                     href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html"
                     title="class in edu.parkside.cs.checklist">Checklist_Contract.Checklist_Queries</a>
         </dt>
         <dd>&nbsp;</dd>
         <dt><span class="strong"><a
-                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#fetchItem(edu.parkside.cs.checklist.Checklist_Item_Row)">fetchItem(Checklist_Item_Row)</a></span>
+                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#fetchItem(edu.parkside.cs.checklist.ChecklistItemRow)">fetchItem(Checklist_Item_Row)</a></span>
             - Static method in class edu.parkside.cs.checklist.<a
                     href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html"
                     title="class in edu.parkside.cs.checklist">Checklist_Contract.Checklist_Item_Queries</a>
         </dt>
         <dd>&nbsp;</dd>
         <dt><span class="strong"><a
-                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#fetchItems(edu.parkside.cs.checklist.Checklist_Row)">fetchItems(Checklist_Row)</a></span>
+                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#fetchItems(edu.parkside.cs.checklist.ChecklistRow)">fetchItems(Checklist_Row)</a></span>
             - Static method in class edu.parkside.cs.checklist.<a
                     href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html"
                     title="class in edu.parkside.cs.checklist">Checklist_Contract.Checklist_Item_Queries</a>
