+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_API_Documentation/index-files/index-7.html
@@ -79,28 +79,28 @@
     <h2 class="title">I</h2>
     <dl>
         <dt><span class="strong"><a
-                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html#insertChecklist(edu.parkside.cs.checklist.Checklist_Row)">insertChecklist(Checklist_Row)</a></span>
+                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html#insertChecklist(edu.parkside.cs.checklist.ChecklistRow)">insertChecklist(Checklist_Row)</a></span>
             - Static method in class edu.parkside.cs.checklist.<a
                     href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Queries.html"
                     title="class in edu.parkside.cs.checklist">Checklist_Contract.Checklist_Queries</a>
         </dt>
         <dd>&nbsp;</dd>
         <dt><span class="strong"><a
-                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Description_Qureries.html#insertDescription(edu.parkside.cs.checklist.Checklist_Item_Row,%20java.lang.String)">insertDescription(Checklist_Item_Row,
+                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Description_Qureries.html#insertDescription(edu.parkside.cs.checklist.ChecklistItemRow,%20java.lang.String)">insertDescription(Checklist_Item_Row,
             String)</a></span> - Static method in class edu.parkside.cs.checklist.<a
                 href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Description_Qureries.html"
                 title="class in edu.parkside.cs.checklist">Checklist_Contract.Checklist_Description_Qureries</a>
         </dt>
         <dd>&nbsp;</dd>
         <dt><span class="strong"><a
-                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#insertItem(edu.parkside.cs.checklist.Checklist_Item_Row)">insertItem(Checklist_Item_Row)</a></span>
+                href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html#insertItem(edu.parkside.cs.checklist.ChecklistItemRow)">insertItem(Checklist_Item_Row)</a></span>
             - Static method in class edu.parkside.cs.checklist.<a
                     href="../edu/parkside/cs/checklist/Checklist_Contract.Checklist_Item_Queries.html"
                     title="class in edu.parkside.cs.checklist">Checklist_Contract.Checklist_Item_Queries</a>
         </dt>
         <dd>&nbsp;</dd>
         <dt><span class="strong"><a
-                href="../edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html#insertItem(edu.parkside.cs.checklist.Checklist_Item_Row,%20java.lang.String)">insertItem(Checklist_Item_Row,
+                href="../edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html#insertItem(edu.parkside.cs.checklist.ChecklistItemRow,%20java.lang.String)">insertItem(Checklist_Item_Row,
             String)</a></span> - Method in class edu.parkside.cs.checklist.<a
                 href="../edu/parkside/cs/checklist/Checklist_Contract_Db_Helper.html"
                 title="class in edu.parkside.cs.checklist">Checklist_Contract_Db_Helper</a></dt>
