+++ app/src/main/java/edu/parkside/cs/checklist/Checklist_Contract.java
@@ -249,7 +249,7 @@ public final class Checklist_Contract {
     /* Inner class that contains the Checklist SQL queries. */
     public static abstract class Checklist_Queries {
 
-        public static final String insertChecklist(Checklist_Row checklist_row) {
+        public static final String insertChecklist(ChecklistRow checklist_row) {
             String query = "INSERT INTO " + Checklist.TABLE_NAME +
                     " VALUES (NULL, " + "\'" + checklist_row.getTitle() + "\'" +
                     ", " + "\'" + checklist_row.getProgress() + "\'" + ")";
@@ -257,13 +257,13 @@ public final class Checklist_Contract {
             return query;
         }
 
-        public static final String fetchChecklist(Checklist_Row checklist_row) {
+        public static final String fetchChecklist(ChecklistRow checklist_row) {
             String query = "SELECT * FROM " + Checklist.TABLE_NAME +
                     " WHERE " + Checklist.COLUMN_NAME_TITLE + " = " + "\'" + checklist_row.getTitle() + "\'";
             return query;
         }
 
-        public static final String updateChecklist(Checklist_Row checklist_row) {
+        public static final String updateChecklist(ChecklistRow checklist_row) {
 
             String query = "UPDATE " + Checklist.TABLE_NAME + " SET " +
                     Checklist.COLUMN_NAME_TITLE + " = " + "\'" + checklist_row.getTitle() + "\'" +
@@ -273,7 +273,7 @@ public final class Checklist_Contract {
             return query;
         }
 
-        public static final String deleteChecklist(Checklist_Row checklist_row) {
+        public static final String deleteChecklist(ChecklistRow checklist_row) {
             String query = "DELETE FROM " + Checklist.TABLE_NAME +
                     " WHERE " + Checklist._ID + " = " + "\'" + checklist_row.getEntryid() + "\'";
             return query;
@@ -304,19 +304,19 @@ public final class Checklist_Contract {
         public static final String ALL_ITEMS = "SELECT * FROM " +
                 Item.TABLE_NAME;
 
-        public static final String fetchItem(Checklist_Item_Row item) {
+        public static final String fetchItem(ChecklistItemRow item) {
             String query = "SELECT * FROM " + Item.TABLE_NAME +
                     " WHERE " + Item.COLUMN_NAME_NAME + " = " + "\'" + item.getName() + "\'";
             return query;
         }
 
-        public static final String fetchItems(Checklist_Row checklist) {
+        public static final String fetchItems(ChecklistRow checklist) {
             String query = "SELECT * FROM " + Item.TABLE_NAME +
                     " WHERE " + Item.COLUMN_NAME_CHECKLIST_ID + " = " + "\'" + checklist.getEntryid() + "\'";
             return query;
         }
 
-        public static final String insertItem(Checklist_Item_Row item) {
+        public static final String insertItem(ChecklistItemRow item) {
             String query = "INSERT INTO " + Item.TABLE_NAME + " (" +
                     Item._ID + ", " + Item.COLUMN_NAME_NAME + ", " +
                     Item.COLUMN_NAME_QTY + ", " + Item.COLUMN_NAME_COMPLETE + ", " +
@@ -330,7 +330,7 @@ public final class Checklist_Contract {
         }
 
 
-        public static final String updateItem(Checklist_Item_Row item) {
+        public static final String updateItem(ChecklistItemRow item) {
             String query = "UPDATE " + Item.TABLE_NAME +
                     " SET " + Item.COLUMN_NAME_NAME + " = " + "\'" + item.getName() + "\'" + ", " +
                     Item.COLUMN_NAME_QTY + " = " + "\'" + item.getQty() + "\', " +
@@ -340,7 +340,7 @@ public final class Checklist_Contract {
             return query;
         }
 
-        public static final String[] deleteItem(Checklist_Item_Row item) {
+        public static final String[] deleteItem(ChecklistItemRow item) {
 
             String query = Checklist_Description_Qureries.deleteDescription(item);
 
@@ -365,7 +365,7 @@ public final class Checklist_Contract {
             return queryString;
         }
 
-        public static final String insertDescription(Checklist_Item_Row item, String description) {
+        public static final String insertDescription(ChecklistItemRow item, String description) {
             String query = "INSERT INTO " + Description.TABLE_NAME + " (" + Description._ID + ", " +
                     Description.COLUMN_NAME_DESCRIPTION + ", " + Description.COLUMN_NAME_ITEM_ID +
                     ") VALUES (NULL, " +
@@ -375,14 +375,14 @@ public final class Checklist_Contract {
             return query;
         }
 
-        public static final String updateDescription(Checklist_Item_Row item, String description) {
+        public static final String updateDescription(ChecklistItemRow item, String description) {
             String query = "UPDATE " + Description.TABLE_NAME +
                     " SET " + Description.COLUMN_NAME_DESCRIPTION + " = " + "\'" + description + "\'" +
                     " WHERE " + Description.COLUMN_NAME_ITEM_ID + " = " + "\'" + item.getEntryid() + "\'";
             return query;
         }
 
-        public static final String deleteDescription(Checklist_Item_Row item) {
+        public static final String deleteDescription(ChecklistItemRow item) {
             String query = "DELETE FROM " + Description.TABLE_NAME
                     + " WHERE " + Description.COLUMN_NAME_ITEM_ID + " = " + "\'" + item.getEntryid() + "\'";
             return query;
