+++ app/src/main/res/layout/activity_checklist_item_listview.xml
@@ -4,7 +4,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="edu.parkside.cs.checklist.Checklist_Item_ListView">
+    tools:context="edu.parkside.cs.checklist.ChecklistItemListView">
 
     <ListView
         android:layout_width="wrap_content"
