+++ app/src/main/res/menu/menu_checklist__item_list_view.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="edu.parkside.cs.checklist.Checklist_Item_ListView">
+    tools:context="edu.parkside.cs.checklist.ChecklistItemListView">
 
     <item android:id="@+id/action_edit"
         android:icon="@drawable/ic_action_edit"
