+++ app/src/main/res/menu/menu_checklist__kit__create.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="edu.parkside.cs.checklist.Checklist_Create">
+    tools:context="edu.parkside.cs.checklist.ChecklistCreate">
     <item android:id="@+id/action_settings" android:title="@string/action_settings"
         android:orderInCategory="100" app:showAsAction="never" />
 </menu>
